requirements.txt
vonage
flask
requests
python-dotenv